Move ResultWidget to own file
